crawler/requirements.txt
scrapy>=2.12.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pydantic>=2.6.0
httpx==0.25.2
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0 